requirements.txt
bottle==0.12.21
bottle-cors-plugin==0.1.9
distlib==0.3.4
fancycompleter==0.9.1
filelock==3.7.1
gunicorn==20.1.0
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
Pygments==2.12.0
pyparsing==3.0.9
pyrepl==0.9.0
pytesseract==0.3.9
six==1.16.0
toml==0.10.2
virtualenv==20.14.1
wmctrl==0.4
